Add initial domoticz role
This adds the install, update and management support for domoticz. Because
the security of this application is slightly in question it is locked down
as far as the systemd builtin MAC capabilities allows.

The idea is to use an nginx ssl proxy that will share the same private
network namespace as the domoticz service. This ensures that domoticz can
never listen on anything but localhost, no matter how much it missbehaves.

Using nginx also means it will be much easier to enable true ssl/tls
certificates, using something like letsencrypt.org.