Add specialized nginx role for domoticz proxy
This nginx setup is used to proxy http and https connections to the
domoticz.service, which only ever listens on localhost inside a
private network namespace.

Nginx runs completly unprivileged by using systemd socket activation.
A simplified overview of how this works.
1. The unit nginx.socket listens on the global namespace.
2. The unit nginx.service is activated, using an ondocumented
   socket-passing mechanism, see:
https://freedesktop.org/wiki/Software/systemd/DaemonSocketActivation/
3. Nginx will proxy all connections to Domoticz, which listens on port
   8080 inside a private net namespace. This is possible, because
   systemd allows nginx.service to join the private namespace of
   domoticz.service

Nginx really wants to act like a traditional Unix daemon, therefore a
lot of configuration files need to be tweaked.

At this time https is not used. First something like a letsencrypt
role must be implemented. A decision must be made if self-signed
certificates should be possible for local https only, or if there
is a simple and sane way of passing the letsencrypt domain
verification without permanently allowing external connections.