Unify nginx configuration
This is done because order is important and when using server_name
the same value cannot be used twice (in different server blocks).

This new configuration enables location/uri based proxying of both
domoticz and ha-bridge. Now it should be possible to connect, using
a fqdn. If avahi is installed and configured that can be used too.

Ensuring avahi is installed may be explicitly added in the future.