Disable private network namespacing
This seemed like a neat idea, but in reality one or even both of
the applications currently supported sometimes require access to
devices on the network.

The current plan is to implement security with firewalls instead.

In the future this may be re-enabled but made optional and off by
default.