Expand systemd based sandboxing
This adds the use of MemoryDenyWriteExecute (where appropriate),
ProtectKernelModules, RestrictRealtime and RestrictAddressFamilies.

Only the last mentioned one is in use on a Raspian based system,
because of the older systemd version. It would be nice to check the
systemd version before templating the unit file, but since systemd
gracefully ignores unknown values this is not neccessary for now.