Add automatic update capabities for domoticz
The automatic updates built in to Domoticz makes insane assumptions,
like having unrestriced passwordless sudo access for restarting
services.

This commit adds a well thought out strategy that adheres to security
best practices as far as possible, domoticz allowing. All updates are
managed by the system, not the application.

A summary of the update strategy:
1. A systemd timer runs daily update checks.
2. Unprivileged user (domoticz) checks for updates, using the same
   method/api domoticz itself uses. If updates are available, files
   are downloaded to /tmp. To not overcomplicate things downloads this
   is also done by the domoticz user.
3. If updates have downloaded successfully, another script is run, now
   under the root user. This script extracts the archive contents and
   restarts the domoticz service.

Less verbose:
  - domoticz-get-update.service
    - runs daily, using domoticz-get-update.timer
    - wants domoticz-update.service
    - before domoticz-update.service
    - executes domoticz-get-update.py
      - checks for update
      - downloads update
    - domoticz-update.service
      - after domoticz-get-update.service
      - only runs if updates downloaded
      - executes domoticz-update.sh
        - extracts files
        - restarts domoticz.service
        - remove archive

This is all managed under systemd. Although the design is flexible and
could be done entirely in cron, with only minor disadvantages.

The third point (3.) theoretically enables most domoticz files to be
owned by root, making them inaccessible to the domoticz service. This
is currently not taken advantage of, due to a lack of testing. Also,
this is made less relevant if most of /opt/domoticz can be made
read-only via namespaces.